+++ packages/web3-errors/src/errors/ens_errors.ts
@@ -22,23 +22,23 @@ import {
 	ERR_ENS_NETWORK_NOT_SYNCED,
 	ERR_ENS_UNSUPPORTED_NETWORK,
 } from '../error_codes';
-import { Web3Error } from '../web3_error_base';
+import { BaseWeb3Error } from '../web3_error_base';
 
-export class ENSCheckInterfaceSupportError extends Web3Error {
+export class ENSCheckInterfaceSupportError extends BaseWeb3Error {
 	public code = ERR_ENS_CHECK_INTERFACE_SUPPORT;
 	public constructor(errorDetails: string) {
 		super(`ENS resolver check interface support error. "${errorDetails}"`);
 	}
 }
 
-export class ENSUnsupportedNetworkError extends Web3Error {
+export class ENSUnsupportedNetworkError extends BaseWeb3Error {
 	public code = ERR_ENS_UNSUPPORTED_NETWORK;
 	public constructor(networkType: string) {
 		super(`ENS is not supported on network ${networkType}`);
 	}
 }
 
-export class ENSNetworkNotSyncedError extends Web3Error {
+export class ENSNetworkNotSyncedError extends BaseWeb3Error {
 	public code = ERR_ENS_NETWORK_NOT_SYNCED;
 	public constructor() {
 		super(`Network not synced`);