+++ packages/web3-eth-contract/src/contract.ts
@@ -27,6 +27,14 @@ import {
 	SendTransactionEvents,
 } from 'web3-eth';
 import {
+	encodeEventSignature,
+	encodeFunctionSignature,
+	isAbiErrorFragment,
+	isAbiEventFragment,
+	isAbiFunctionFragment,
+	jsonInterfaceMethodToString,
+} from 'web3-eth-abi';
+import {
 	AbiConstructorFragment,
 	AbiErrorFragment,
 	AbiEventFragment,
@@ -39,21 +47,13 @@ import {
 	ContractMethod,
 	ContractMethodInputParameters,
 	ContractMethodOutputParameters,
-	encodeEventSignature,
-	encodeFunctionSignature,
-	FilterAbis,
-	isAbiErrorFragment,
-	isAbiEventFragment,
-	isAbiFunctionFragment,
-	jsonInterfaceMethodToString,
-} from 'web3-eth-abi';
-import {
 	Address,
 	BlockNumberOrTag,
 	BlockTags,
 	Bytes,
 	EthExecutionAPI,
 	Filter,
+	FilterAbis,
 	HexString,
 	LogsInput,
 	Mutable,