+++ packages/web3-utils/src/objects.ts
@@ -15,10 +15,8 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+import { TypedArray } from 'web3-types';
 import { isNullish } from 'web3-validator';
-
-// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-export const TypedArray = Object.getPrototypeOf(Uint8Array);
 
 const isIterable = (item: unknown): item is Record<string, unknown> =>
 	typeof item === 'object' &&