+++ packages/web3-utils/src/web3_deferred_promise.ts
@@ -16,7 +16,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 import { OperationTimeoutError } from 'web3-errors';
-import { Web3DeferredPromise as Web3DeferredPromiseInterface } from 'web3-types';
+import { Web3DeferredPromiseInterface } from 'web3-types';
 
 export class Web3DeferredPromise<T> implements Promise<T>, Web3DeferredPromiseInterface<T> {
 	// public tag to treat object as promise by different libs