+++ packages/web3-providers-http/CHANGELOG.md
@@ -60,3 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   web3.js dependencies (#5757)
 
 ## [Unreleased]
+
+### Added
+
+-   Added named export for `HttpProvider` (#5771)